Zygote: Additional whitelisting for legacy devices.
On M and below, we provide a blanket whitelist for all files under
"/vendor/zygote_whitelist". This path is whitelisted purely to allow
this patch to be applied easily on legacy devices and configurations.

Note that this does not amount to a loosening of our security policy
because whitelisted files are reopened anyway.

Bug: 32691930
Test: manual

(cherry-picked from commit 5e2f7c6229d7191183888d685b57a7d0a2835fce)

Change-Id: I12a3f0d84e3b7454e77f917b71960cd81e2309e3